refactor: [FatSecretToken] FatSecretToken 클래스를 fatsecret 패키지에 속하도록 변경
