+++ src/main/java/com/hotnerds/common/ScheduledTasks.java
@@ -1,5 +1,6 @@
 package com.hotnerds.common;
 
+import com.hotnerds.fatsecret.FatSecretToken;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
